feat: add dedicated pending song manager page
